refactor: change name `wa-worker` to `notification-worker`
