+++ internal/notification-worker/core/service.go
