+++ internal/notification-worker/driven/publisher/wa/wa_test.go
@@ -4,9 +4,9 @@ import (
 	"context"
 	"testing"
 
+	publisher "github.com/ghazlabs/idn-remote-entry/internal/notification-worker/driven/publisher/wa"
 	"github.com/ghazlabs/idn-remote-entry/internal/shared/core"
 	"github.com/ghazlabs/idn-remote-entry/internal/testutil"
-	publisher "github.com/ghazlabs/idn-remote-entry/internal/wa-worker/driven/publisher/wa"
 	"github.com/go-resty/resty/v2"
 	"github.com/riandyrn/go-env"
 	"github.com/stretchr/testify/require"
